+++ CHANGELOG.md
@@ -36,6 +36,8 @@ the release.
   ([#1836](https://github.com/open-telemetry/opentelemetry-demo/pull/1836))
 * [otel-collector] rename otelcol to otel-collector
   ([#1841](https://github.com/open-telemetry/opentelemetry-demo/pull/1841))
+* [shipping] rename shippingservice to accounting
+  ([#1827](https://github.com/open-telemetry/opentelemetry-demo/pull/1842))
 
 ## 1.12.0